+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file. This project adheres
 to [Semantic Versioning](http://semver.org/). For change log format, use [Keep a Changelog](http://keepachangelog.com/).
 
+## [v2.0.0-beta4] - 2022-05-17
+
+### Changed
+
+- Markdown parser use `FrontMatterExtension` of `league/commonmark` package to get metadata
+- Markdown parser accept `ConverterInterface` instead of `EnvironmentInterface`
+
 ## [v2.0.0-beta3] - 2022-05-17
 
 ### Added